Rename iris/device/ to iris/mem/, unify Context class names
Move device-side remote memory access primitives from iris/device/ to
iris/mem/ to better reflect what the module provides. Unify class names
across backends: DeviceContext → Context, IrisDeviceCtx → Context,
DeviceTracing → Tracing, GluonDeviceTracing → Tracing.

iris/device/ files become thin backward-compat shims that re-export from
iris.mem.*. The __translate dunder is explicitly re-exported since star
imports exclude dunder names. All internal imports updated; examples
left unchanged to exercise the shim path.
