+++ iris/device/__init__.py
@@ -1,3 +1,5 @@
 # SPDX-License-Identifier: MIT
 # Copyright (c) 2025-2026 Advanced Micro Devices, Inc. All rights reserved.
-"""Iris device-side code (JIT kernels)."""
+"""Backward compat shim — canonical code moved to iris.mem."""
+
+from iris.mem import *  # noqa: F401,F403
