+++ iris/device/gluon/__init__.py
@@ -1,6 +1,5 @@
 # SPDX-License-Identifier: MIT
 # Copyright (c) 2025-2026 Advanced Micro Devices, Inc. All rights reserved.
-"""Gluon device-side context and tracing."""
+"""Backward compat shim — canonical code moved to iris.mem.gluon."""
 
-from .context import IrisDeviceCtx  # noqa: F401
-from .tracing import GluonDeviceTracing  # noqa: F401
+from iris.mem.gluon import *  # noqa: F401,F403
