+++ iris/x/gather.py
@@ -14,7 +14,7 @@ Use `gather` when you want to consume the tile immediately without materializing
 import triton
 import triton.language as tl
 import iris
-from iris.device.triton.context import DeviceContext
+from iris.mem.triton.context import Context as DeviceContext
 from .core import Tile, TensorView
