+++ iris/x/reduce_scatter.py
@@ -10,7 +10,7 @@ Reduces tiles from all ranks and stores the result only to the assigned rank.
 import triton
 import triton.language as tl
 import iris
-from iris.device.triton.context import DeviceContext
+from iris.mem.triton.context import Context as DeviceContext
 from .core import Tile, TensorView
